+++ demisto_sdk/commands/common/update_id_set.py
@@ -1894,9 +1894,9 @@ def has_duplicate(id_set_subset_list, id_to_check, object_type=None, print_logs=
             dict2_from_version <= dict1_from_version < dict2_to_version,  # will catch (C, B), (B, A), (C, A)
             dict2_from_version < dict1_to_version <= dict2_to_version,  # will catch (C, B), (C, A)
         ]):
-            print_warning('The following {} have the same ID ({}) and their versions overlap: '
-                          '1) "{}-{}", 2) "{}-{}".'.format(object_type, id_to_check, dict1_from_version,
-                                                           dict1_to_version, dict2_from_version, dict2_to_version))
+            print_warning(f'The following {object_type} have the same ID ({id_to_check}) and their versions overlap: '
+                          f'1) "{dict1_from_version}-{dict1_to_version}", '
+                          f'2) "{dict2_from_version}-{dict2_to_version}".')
             return True
 
         if print_logs and dict1.get('name') != dict2.get('name'):