+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Fixed an issue where the **postman code-gen** command generated double dots for context outputs when it was not needed.
 * Fixed an issue where there **validate** command on release notes file crashed when author image was added or modified.
 * Added input handling when running **find-dependencies**, replacing string manipulations.
+* Fixed an issue where the **validate** command did not handle multiple playbooks with the same name in the id_set.
 * Added support for GitLab repositories in **validate**
 
 # 1.5.0